Working URE Real Estate
The URE searching is working and loading into the database, but had to change the KSL search to use selenium. The KSL worked on one listing but was throwing errors after that. The searchKLS function is working but the getKSLDetails seems to be off. Going to pause that work and work on getting all the URE in a csv to load into astro.